# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ksuid": {
            "post": {
                "description": "Create a new user with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create a user",
                "parameters": [
                    {
                        "description": "User object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User Created",
                        "schema": {
                            "$ref": "#/definitions/models.UserInput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/ksuid/{id}": {
            "get": {
                "description": "Get a user by their ID or username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get a single user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "user_name",
                        "in": "path"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "User Found",
                        "schema": {
                            "$ref": "#/definitions/models.UserInput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a user's information by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Updated",
                        "schema": {
                            "$ref": "#/definitions/models.UserInput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a user by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/ksuids": {
            "get": {
                "description": "Get a list of users, with optional search, pagination, and limit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get multiple users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search Term",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Users Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserPaging"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.UserInput": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department is optional.",
                    "type": "string"
                },
                "email": {
                    "description": "Email is required when creating a new user.",
                    "type": "string",
                    "maxLength": 255
                },
                "first_name": {
                    "description": "FirstName is required when creating a new user.",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "id": {
                    "description": "HashId is the public identifier for the user (UUID).\nFor create operations, this might be generated internally.",
                    "type": "string"
                },
                "last_name": {
                    "description": "LastName is required when creating a new user.",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "user_name": {
                    "description": "UserName is required when creating a new user.",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 5
                }
            }
        },
        "models.UserPaging": {
            "description": "UserDTOPaging",
            "type": "object",
            "properties": {
                "page": {
                    "description": "The current page number",
                    "type": "integer"
                },
                "page_count": {
                    "description": "The total number of items available",
                    "type": "integer"
                },
                "page_size": {
                    "description": "The number of items per page",
                    "type": "integer"
                },
                "users": {
                    "description": "A list of users",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserInput"
                    }
                }
            }
        }
    }
}